+++ Assets/Workspace/HSD/Data/Skill.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 10b8fd17a03a02b4ebd760e47783bf17
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
