+++ Assets/Workspace/HSD/Data/Skill/RangedSkill.asset
@@ -1,0 +1,28 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 982e89bf473cf30439e6ae14c18babb9, type: 3}
+  m_Name: RangedSkill
+  m_EditorClassIdentifier: 
+  ManaCost: 10
+  AttackPointOffset: {x: 1, y: 1}
+  AttackPower: 0.2
+  DamageType: 0
+  Priority: 2
+  SizeOrRadius: 5
+  BoxSize: {x: 1, y: 1}
+  Angle: 90
+  MaxCount: 1
+  SearchCount: 1
+  Fov: 45
+  _projectilePrefab: {fileID: 4945829589483694160, guid: eb832e44d39787e4aa23b2f829ac6593,
+    type: 3}
+  projectileSpeed: 10
