+++ Assets/Workspace/HSD/Data/Skill/RangedSkill.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aaed2cefc25c1d845b231a60a11ec433
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
