+++ Assets/Workspace/HSD/Scripts/Unit/FSM/States/States.cs
@@ -167,8 +167,7 @@ public class SkillState : AnimationFinishedState
 
     public override void Enter()
     {
-        base.Enter();
-        
+        base.Enter();        
     }
 
     public override void Exit()
