+++ Assets/Workspace/HSD/Scripts/Unit/Skill/AttackSkill/AttackSkill.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2b47ad94328a68408fe945d1dd53a23
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
