+++ Assets/Workspace/HSD/Scripts/Unit/Skill/AttackSkill/RangedSkill.cs
@@ -1,0 +1,22 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "RangedSkill", menuName = "Data/Unit/Skill/Ranged")]
+public class RangedSkill : AttackSkill
+{
+    [SerializeField] GameObject _projectilePrefab;
+    [SerializeField] float projectileSpeed = 10f;
+
+    public override void Active(IAttacker attacker)
+    {
+        Vector2 offset = AttackPointOffset;
+        offset.x *= attacker.GetTransform().GetFacingDir();
+
+        Vector2 spawnPoint = (Vector2)attacker.GetTransform().position + offset;
+
+        GameObject obj = Instantiate(_projectilePrefab, spawnPoint, Quaternion.identity);
+        Projectile projectile = ComponentProvider.Get<Projectile>(obj);
+        projectile.Init(GetTargetSingle(attacker).transform , attacker.GetStatusController(), AttackPower, DamageType, attacker.TargetLayer, projectileSpeed);
+    }
+}
