+++ Assets/Workspace/HSD/Scripts/Unit/Skill/AttackSkill/RangedSkill.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 982e89bf473cf30439e6ae14c18babb9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
