+++ Assets/Workspace/HSD/Test.unity
@@ -997,6 +997,7 @@ MonoBehaviour:
   _enemyController: {fileID: 859023221}
   _testDatas:
   - {fileID: 11400000, guid: 0213d327ee461674dbd620f59eacfcdb, type: 2}
+  - {fileID: 11400000, guid: 16de7f9473f1a3940bab2394a06f970d, type: 2}
   _upgradeNeedCount: 3
 --- !u!4 &651601983
 Transform:
